:recycle: refactor(api) extract proOrganization creation
